+++ Escon.SisctNET.Web/Controllers/TaxAnexoController.cs
@@ -230,10 +230,10 @@ namespace Escon.SisctNET.Web.Controllers
 
                     stream.Close();
 
-                    if (comp.AnnexId.Equals(1))
+                    if (comp.AnnexId.Equals((long)1))
                     {
                         var notes = _noteService.FindByNotes(companyid, year, month);
-                        var products = _productNoteService.FindByProductsType(notes, Model.TypeTaxation.Nenhum).Where(_ => _.TaxationTypeId.Equals(1) && _.TaxationTypeId.Equals(3) && _.Incentivo.Equals(false)).ToList();
+                        var products = _productNoteService.FindByProductsType(notes, Model.TypeTaxation.Nenhum).Where(_ => _.TaxationTypeId.Equals((long)1) && _.TaxationTypeId.Equals((long)3) && _.Incentivo.Equals(false)).ToList();
 
                         decimal baseCalcCompraInterestadual4 = products.Where(_ => _.Picms.Equals(4)).Sum(_ => _.Vbasecalc),
                                 baseCalcCompraInterestadual7 = products.Where(_ => _.Picms.Equals(7)).Sum(_ => _.Vbasecalc), 
@@ -413,7 +413,7 @@ namespace Escon.SisctNET.Web.Controllers
                 }
                 else
                 {
-                    if (comp.AnnexId.Equals(1))
+                    if (comp.AnnexId.Equals((long)1))
                     {
                         exitNotes = importXml.NFeAll(directoryNfeExit);
